14-02-2019

+++ app/src/main/java/com/blacksite/scarystories/model/ParagraphScene.kt
@@ -3,7 +3,7 @@ package com.blacksite.scarystories.model
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class ParagraphScene(@SerializedName("position")var position:Int,
+data class ParagraphScene(@SerializedName("position")var position:Double,
                           @SerializedName("playTime")var playTime:Int,
                           @SerializedName("duration")var duration:Int,
                           @SerializedName("paragraph")var paragraph: Paragraph): Serializable {
